разбил проект по blueprint. мелкий рефакторинг. проверка номера участка при регистрации, добавил папку для актов. починил скачивание актов из лк. как бы применить декоратор админа для доступа ко всем лк?
