+++ webapp/config.py
@@ -1,3 +1,4 @@
+from datetime import timedelta
 import os
 from dotenv import load_dotenv
 load_dotenv()
@@ -7,6 +8,7 @@ SECRET_KEY = os.getenv('SECRET_KEY')
 WTF_CSRF_SECRET_KEY = os.getenv('WTF_CSRF_SECRET_KEY')
 basedir = os.path.abspath(os.path.dirname(__file__))
 SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, '..', 'webapp.db')
+REMEMBER_COOKIE_DURATION = timedelta(days=30)
 
 
 # Константы для функции парсинга
