+++ webapp/templates/contacts.html
@@ -9,7 +9,7 @@
           </a>
         </div>
             <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-              <li><a href="{{ url_for('index') }}" class="nav-link px-2 link-secondary text-start ">Главная страница</a></li>
+              <li><a href="{{ url_for('news.index') }}" class="nav-link px-2 link-secondary text-start ">Главная страница</a></li>
             </ul>
         <div class="col-md-3 text-end">
