+++ webapp/templates/index.html
@@ -1,34 +1,38 @@
 {% extends "base.html" %}
 {% block title %}{{ page_title }}{% endblock %}
 {% block head %}
-      <div class="container">
-        <header class="d-flex flex-wrap align-items-center justify-content-center justify-content-md-between py-1 mb-2 border-bottom">
-          <div class="col-md-3 mb-2 mb-md-0">
-            <a href="{{ url_for('index') }}" class="d-inline-flex link-body-emphasis text-decoration-none">
-              <img class="head_img" src="{{ url_for('static', filename='img/icon.png') }}" alt='pic'>
+<div class="container">
+    <header class="d-flex flex-wrap align-items-center justify-content-center justify-content-md-between py-1 mb-2 border-bottom">
+        <div class="col-md-3 mb-2 mb-md-0">
+            <a href="{{ url_for('news.index') }}" class="d-inline-flex link-body-emphasis text-decoration-none">
+                <img class="head_img" src="{{ url_for('static', filename='img/icon.png') }}" alt='pic'>
             </a>
-          </div>
-          <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-          </ul>
-          <div class="col-md-3 text-end">
-            <a class="btn btn-outline-primary me-2" href="{{ url_for('login') }}" role="button">Войти</a>
-            <a class="btn btn-primary" href="{{ url_for('registration') }}" role="button">Зарегистрироваться</a>
-            </div>
-        </header>
         </div>
+        <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
+            <li>
+            <li><a href="{{ url_for('news.index') }}" class="nav-link px-2 link-secondary text-start ">Сайт ТСН
+                "Восток"</a></li>
+            </li>
+        </ul>
+        <div class="col-md-3 text-end">
+            <a class="btn btn-outline-primary me-2" href="{{ url_for('user.login') }}" role="button">Войти</a>
+            <a class="btn btn-primary" href="{{ url_for('user.registration') }}" role="button">Зарегистрироваться</a>
+        </div>
+    </header>
+</div>
 {% endblock %}
 {% block body %}
-              <div class="container">
-                 {% with messages = get_flashed_messages() %}
-                    {% if messages %}
-                        <div class="alert alert-warning" role="alert">
-                            {% for message in messages %}
-                                {{ message }}<br>
-                            {% endfor %}
-                        </div>
-                    {% endif %}
-                 {% endwith %}
-              </div>
+<div class="container">
+    {% with messages = get_flashed_messages() %}
+    {% if messages %}
+    <div class="alert alert-warning" role="alert">
+        {% for message in messages %}
+        {{ message }}<br>
+        {% endfor %}
+    </div>
+    {% endif %}
+    {% endwith %}
+</div>
 <!--    <div class="container">-->
 <!--        <div class="row">-->
 <!--          <div class="col">-->
