+++ webapp/templates/lk_page.html
@@ -5,8 +5,7 @@
 <div class="container">
     <div class="flex-container">
         <div class="first">
-            <a href="/{{ area }}.xls" style="color: black" download>Скачать акт сверки</a>
-
+            <a href="/static/acts/{{ area }}.xls" style="color: black" download>Скачать акт сверки</a>
         </div>
         <div class="second">
             <p style="text-align: center">Добро пожаловать в личный кабинет участка № {{ area }}</p>
