+++ webapp/templates/registration.html
@@ -12,7 +12,7 @@
             </div>
         {% endif %}
      {% endwith %}
-    <form action="{{ url_for('reg_processing') }}" method="POST">
+    <form action="{{ url_for('user.reg_processing') }}" method="POST">
 
         <h1 class="h3 mb-3 fw-normal text-center">
             <a href="/" > <img class="head_img" src="{{ url_for('static', filename='img/icon.png') }}" alt='pic'></a>
@@ -38,8 +38,7 @@
                 {{ form.confirm(mod="password1") }} {{ form.confirm.label }}
 
         </div>
-
-                   {{ form.submit() }}
+                {{ form.submit() }}
         <p class="mt-2 mb-3 text-body-secondary">&copy;2024</p>
     </form>
 </main>
