+++ webapp/user/models.py
@@ -1,10 +1,6 @@
-import datetime
-from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin
-from sqlalchemy.orm import Mapped, mapped_column
 from werkzeug.security import generate_password_hash, check_password_hash
-
-db = SQLAlchemy()
+from webapp.db import db
 
 
 class User(db.Model, UserMixin):
@@ -29,32 +25,9 @@ class User(db.Model, UserMixin):
     def check_password(self, password):
         return check_password_hash(self.password, password)
 
-
     def __repr__(self):
         return f'Пользователь с участка {self.area_number} с id={self.id}'
 
     @property
     def is_admin(self):
         return self.role == 'admin'
-
-
-class News(db.Model):
-    __tablename__ = 'news_table'
-    __table_args__ = {'comment': 'Таблица с новостями на главной странице'}
-    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
-    published: Mapped[datetime.datetime] = mapped_column(nullable=False)
-    text: Mapped[str] = mapped_column(nullable=True)
-
-    def __repr__(self):
-        return '<News {} {}>'.format(self.id, self.published)
-
-
-class FinancialData(db.Model):
-    __tablename__ = 'financial_data'
-    __table_args__ = {'comment': 'Общая финансовая таблица, загружаемая извне в ЛК у админа'}
-    id: Mapped[int] = mapped_column(autoincrement=True, nullable=False)
-    area_number: Mapped[int] = mapped_column(primary_key=True, nullable=False)
-    member_fee: Mapped[int] = mapped_column(nullable=True)
-    targeted_fee: Mapped[int] = mapped_column(nullable=True)
-    electricity_payments: Mapped[int] = mapped_column(nullable=True)
-    published: Mapped[datetime.date] = mapped_column(nullable=False)
